refactor(ob1)!: drop the top-level OB1 compat shim modules (#235)
The same OB1 legacy name was reachable three ways — the unprefixed top-level
re-exports (openbadgeslib.Signer), openbadgeslib.ob1.Signer, and the top-level
openbadgeslib.badge/signer/verifier shim modules — each with its own warning,
and the shims imported ob1 eagerly just to warn. With the name table now
single-sourced (previous commit), retire the redundant shim modules.

- Delete openbadgeslib/badge.py, openbadgeslib/signer.py and
  openbadgeslib/verifier.py. OpenBadges 1.0 stays fully supported and reachable
  via openbadgeslib.ob1.* (from openbadgeslib.ob1 import Signer) and the
  unprefixed openbadgeslib.Signer / .Badge / … re-exports, both of which still
  emit the steering DeprecationWarning.
- Repoint the test suite (conftest and the ob1/cli tests) from the removed
  shim paths to openbadgeslib.ob1.<module> (the real leaf, which is silent for
  internal use). Replace the shim-warns deprecation test with one asserting the
  three modules are gone (ImportError).
- Update the wiki (Python-API-OB1, OB2-vs-OB3, Stable-API, Contributing) to
  describe OB1 via ob1.* / the unprefixed names and note the shim removal.

Full suite (1277) + flake8 + mypy --strict green.

BREAKING CHANGE: `from openbadgeslib.badge import ...` and the .signer/
.verifier module paths no longer resolve; use openbadgeslib.ob1.<module> or the
unprefixed top-level names. Lands in 4.0.

Closes #235
