--- openbadgeslib/badge.py
@@ -1,21 +1,0 @@
-"""OpenBadges 1.0 (legacy) badge objects — compatibility shim, re-exports from openbadgeslib.ob1."""
-import warnings
-
-from .ob1.badge import (
-    BadgeStatus, BadgeImgType, BadgeType,
-    Assertion, Badge, BadgeSigned,
-    extract_svg_assertion, extract_png_assertion,
-)
-
-warnings.warn(
-    'openbadgeslib.badge is a legacy OpenBadges 1.0 compatibility shim; prefer '
-    'openbadgeslib.ob2 (strict OB 2.0) or openbadgeslib.ob3 (OB 3.0) for new '
-    'work. OpenBadges 1.0 remains supported (no removal planned). See the '
-    '"OpenBadges 1.0 lifecycle" wiki page.',
-    DeprecationWarning, stacklevel=2)
-
-__all__ = [
-    'BadgeStatus', 'BadgeImgType', 'BadgeType',
-    'Assertion', 'Badge', 'BadgeSigned',
-    'extract_svg_assertion', 'extract_png_assertion',
-]
