+++ wiki/Contributing.md
@@ -130,7 +130,7 @@ tests/
     fixtures/            schema + test-vector data files
 ```
 
-Note: `baking.py` is a single shared top-level module (`openbadgeslib.baking`) that embeds/extracts the token in the SVG/PNG carrier; the OB1, OB2 and OB3 signers/verifiers all call it. The legacy OB 1.0 implementation lives in `ob1/` (the top-level `badge.py`, `signer.py`, and `verifier.py` are **legacy** compatibility shims that re-export it so imports like `from openbadgeslib.badge import Badge` keep working, now with a `DeprecationWarning` steering new work to OB2/OB3 — OB 1.0 stays supported, see [[Python API OB1]]); the strict OB 2.0 implementation lives in `ob2/` (`from openbadgeslib.ob2 import OB2Signer, OB2Verifier`). For how the generations differ see [[OB2 vs OB3]], and for the public interfaces see [[Python API OB1]], [[Python API OB2]] and [[Python API OB3]].
+Note: `baking.py` is a single shared top-level module (`openbadgeslib.baking`) that embeds/extracts the token in the SVG/PNG carrier; the OB1, OB2 and OB3 signers/verifiers all call it. The legacy OB 1.0 implementation lives in `ob1/` (reached via `from openbadgeslib.ob1 import Signer` or the unprefixed `openbadgeslib.Signer`, both of which emit a `DeprecationWarning` steering new work to OB2/OB3 — OB 1.0 stays supported; the old top-level `badge.py`/`signer.py`/`verifier.py` shim modules were dropped in v4, see [[Python API OB1]]); the strict OB 2.0 implementation lives in `ob2/` (`from openbadgeslib.ob2 import OB2Signer, OB2Verifier`). For how the generations differ see [[OB2 vs OB3]], and for the public interfaces see [[Python API OB1]], [[Python API OB2]] and [[Python API OB3]].
 
 ## Continuous integration
